Refactored code into packages and fixed IntelliJ run issues
