--- src/main/java/Deadline.java
@@ -1,23 +1,0 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-
-public class Deadline extends Task {
-    protected LocalDate by;
-
-    public Deadline(String description, String by) {
-        super(description, TaskType.DEADLINE);
-        try {
-            this.by = LocalDate.parse(by);
-        } catch (DateTimeParseException e) {
-            throw new IllegalArgumentException("Invalid date format! Use yyyy-MM-dd.");
-        }
-
-    }
-
-    @Override
-    public String toString() {
-        return super.toString() + " (by: "
-                + by.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
-    }
-}
