--- src/main/java/Event.java
@@ -1,25 +1,0 @@
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-
-public class Event extends Task {
-    protected LocalDate from;
-    protected LocalDate to;
-
-    public Event(String description, String from, String to) {
-        super(description, TaskType.EVENT);
-        try {
-            this.from = LocalDate.parse(from);
-            this.to = LocalDate.parse(to);
-        } catch (DateTimeParseException e) {
-            throw new IllegalArgumentException("Invalid date format! Use yyyy-MM-dd.");
-        }
-    }
-
-    @Override
-    public String toString() {
-        return "[E]" + super.toString() + " (from: "
-                + from.format(DateTimeFormatter.ofPattern("MMM dd yyyy"))
-                + " to: " + to.format(DateTimeFormatter.ofPattern("MMM dd yyyy")) + ")";
-    }
-}
