+++ src/main/java/shin/Parser.java
@@ -1,3 +1,5 @@
+package shin;
+
 public class Parser {
     public static String[] parse(String input) {
         return input.trim().split(" ", 2);  // Splits command and arguments
