+++ src/main/java/shin/exception/ShinException.java
@@ -1,3 +1,5 @@
+package shin.exception;
+
 public class ShinException extends Exception {
     public ShinException(String message) {
         super(message);
