+++ src/main/java/shin/task/Deadline.java
@@ -1,0 +1,23 @@
+package shin.task;
+
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    protected LocalDate by;
+
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = LocalDate.parse(by);
+    }
+
+    public LocalDate getBy() {   // ✅ Add this getter
+        return by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: "
+                + by.format(DateTimeFormatter.ofPattern("MMM d yyyy")) + ")";
+    }
+}
