+++ src/main/java/shin/task/Task.java
@@ -1,0 +1,27 @@
+package shin.task;
+
+public class Task {
+    protected String description;
+    protected boolean isDone;
+
+    public Task(String description) {
+        this.description = description;
+        this.isDone = false;
+    }
+
+    public String getDescription() {   // ✅ Add this getter
+        return description;
+    }
+
+    public boolean isDone() {   // ✅ Add this getter
+        return isDone;
+    }
+
+    public void markAsDone() {
+        isDone = true;
+    }
+
+    public void markAsNotDone() {
+        isDone = false;
+    }
+}
