+++ src/main/java/shin/task/TaskType.java
@@ -1,3 +1,4 @@
+package shin.task;
 public enum TaskType {
     TODO, DEADLINE, EVENT
 }
