+++ src/main/java/shin/task/Todo.java
@@ -1,7 +1,11 @@
+package shin.task;
+
+import shin.task.Task;
+
 public class Todo extends Task {
 
     public Todo(String description) {
-        super(description, TaskType.TODO);
+        super(description);
     }
 
     @Override
